Move config file logic into config.go
