+++ config_test.go
@@ -2,6 +2,9 @@ package main
 
 import (
 	"encoding/json"
+	"io/ioutil"
+	"os"
+	"path/filepath"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -18,4 +21,140 @@ func TestLinterConfigUnmarshalJSON(t *testing.T) {
 	require.NoError(t, err)
 	assert.Equal(t, "/bin/custom", config.Command)
 	assert.Equal(t, functionName(partitionPathsAsDirectories), functionName(config.PartitionStrategy))
+}
+
+func TestFindDefaultConfigFile(t *testing.T) {
+	tmpdir, cleanup := setupTempDir(t)
+	defer cleanup()
+
+	mkDir(t, tmpdir, "contains")
+	mkDir(t, tmpdir, "contains", "foo")
+	mkDir(t, tmpdir, "contains", "foo", "bar")
+	mkDir(t, tmpdir, "contains", "double")
+	mkDir(t, tmpdir, "lacks")
+
+	mkConfigFile(t, filepath.Join(tmpdir, "contains"), defaultConfigPath)
+	mkConfigFile(t, filepath.Join(tmpdir, "contains", "double"), defaultConfigPath)
+
+	var testcases = []struct {
+		dir      string
+		expected string
+		found    bool
+	}{
+		{
+			dir:      tmpdir,
+			expected: "",
+			found:    false,
+		},
+		{
+			dir:      filepath.Join(tmpdir, "contains"),
+			expected: filepath.Join(tmpdir, "contains", defaultConfigPath),
+			found:    true,
+		},
+		{
+			dir:      filepath.Join(tmpdir, "contains", "foo"),
+			expected: filepath.Join(tmpdir, "contains", defaultConfigPath),
+			found:    true,
+		},
+		{
+			dir:      filepath.Join(tmpdir, "contains", "foo", "bar"),
+			expected: filepath.Join(tmpdir, "contains", defaultConfigPath),
+			found:    true,
+		},
+		{
+			dir:      filepath.Join(tmpdir, "contains", "double"),
+			expected: filepath.Join(tmpdir, "contains", "double", defaultConfigPath),
+			found:    true,
+		},
+		{
+			dir:      filepath.Join(tmpdir, "lacks"),
+			expected: "",
+			found:    false,
+		},
+	}
+
+	for _, testcase := range testcases {
+		require.NoError(t, os.Chdir(testcase.dir))
+		configFile, found, err := findDefaultConfigFile()
+		assert.Equal(t, testcase.expected, configFile)
+		assert.Equal(t, testcase.found, found)
+		assert.NoError(t, err)
+	}
+}
+
+func TestFindDefaultConfigFileWithHome(t *testing.T) {
+	tmpdir, cleanup := setupTempDir(t)
+	defer cleanup()
+
+	homedir, homecleanup := setupTempHomeDir(t, tmpdir)
+	defer homecleanup()
+
+	mkConfigFile(t, homedir, defaultConfigPath)
+
+	mkDir(t, tmpdir, "contains")
+	mkDir(t, tmpdir, "contains", "foo")
+	mkDir(t, tmpdir, "contains", "foo", "bar")
+	mkDir(t, tmpdir, "contains", "double")
+	mkDir(t, tmpdir, "lacks")
+
+	mkConfigFile(t, filepath.Join(tmpdir, "contains"), defaultConfigPath)
+	mkConfigFile(t, filepath.Join(tmpdir, "contains", "double"), defaultConfigPath)
+
+	var testcases = []struct {
+		dir      string
+		expected string
+	}{
+		{
+			dir:      tmpdir,
+			expected: filepath.Join(homedir, defaultConfigPath),
+		},
+		{
+			dir:      filepath.Join(tmpdir, "contains"),
+			expected: filepath.Join(tmpdir, "contains", defaultConfigPath),
+		},
+		{
+			dir:      filepath.Join(tmpdir, "contains", "foo"),
+			expected: filepath.Join(tmpdir, "contains", defaultConfigPath),
+		},
+		{
+			dir:      filepath.Join(tmpdir, "contains", "foo", "bar"),
+			expected: filepath.Join(tmpdir, "contains", defaultConfigPath),
+		},
+		{
+			dir:      filepath.Join(tmpdir, "contains", "double"),
+			expected: filepath.Join(tmpdir, "contains", "double", defaultConfigPath),
+		},
+		{
+			dir:      filepath.Join(tmpdir, "lacks"),
+			expected: filepath.Join(homedir, defaultConfigPath),
+		},
+	}
+
+	for _, testcase := range testcases {
+		require.NoError(t, os.Chdir(testcase.dir))
+		configFile, found, err := findDefaultConfigFile()
+		assert.Equal(t, testcase.expected, configFile)
+		assert.True(t, found)
+		assert.NoError(t, err)
+	}
+}
+
+func setupTempHomeDir(t *testing.T, tmpdir string) (string, func()) {
+	homeDir := filepath.Join(tmpdir, "test-home")
+	mkDir(t, homeDir)
+
+	cachedFunc := getHomeDir
+	getHomeDir = func() (string, error) {
+		return homeDir, nil
+	}
+
+	return homeDir, func() {
+		getHomeDir = cachedFunc
+	}
+}
+
+func mkConfigFile(t *testing.T, path string, filename string) {
+	content := []byte("{}")
+	err := ioutil.WriteFile(filepath.Join(path, filename), content, 0644)
+	require.NoError(t, err)
 }
