+++ main.go
@@ -108,72 +108,6 @@ func loadConfig(app *kingpin.Application, element *kingpin.ParseElement, ctx *ki
 	return loadConfigFile(*element.Value)
 }
 
-func loadConfigFile(filename string) error {
-	r, err := os.Open(filename)
-	if err != nil {
-		return err
-	}
-	defer r.Close() // nolint: errcheck
-	err = json.NewDecoder(r).Decode(config)
-	if err != nil {
-		return err
-	}
-	for _, disable := range config.Disable {
-		for i, enable := range config.Enable {
-			if enable == disable {
-				config.Enable = append(config.Enable[:i], config.Enable[i+1:]...)
-				break
-			}
-		}
-	}
-	return err
-}
-
-func findDefaultConfigFile() (fullPath string, found bool, err error) {
-	prevPath := ""
-	dirPath, err := os.Getwd()
-	if err != nil {
-		return "", false, err
-	}
-
-	for dirPath != prevPath {
-		var fullPath string
-		var found bool
-		fullPath, found, err = findConfigFileInDir(dirPath)
-		if err != nil || found {
-			return fullPath, found, err
-		}
-		prevPath, dirPath = dirPath, filepath.Dir(dirPath)
-	}
-
-	homeDir, err := getHomeDir()
-	if err != nil {
-		return "", false, err
-	}
-
-	return findConfigFileInDir(homeDir)
-}
-
-func findConfigFileInDir(dirPath string) (fullPath string, found bool, err error) {
-	fullPath = filepath.Join(dirPath, defaultConfigPath)
-	if _, err := os.Stat(fullPath); err != nil {
-		if os.IsNotExist(err) {
-			return "", false, nil
-		}
-		return "", false, err
-	}
-
-	return fullPath, true, nil
-}
-
-var getHomeDir = func() (string, error) {
-	u, err := user.Current()
-	if err != nil {
-		return "", err
-	}
-	return u.HomeDir, nil
-}
-
 func disableAction(app *kingpin.Application, element *kingpin.ParseElement, ctx *kingpin.ParseContext) error {
 	out := []string{}
 	for _, linter := range config.Enable {
